agent_platform/requirements.txt
# ── Core Framework ────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
gunicorn>=22.0.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# ── LangChain / LangGraph ────────────────────────────────────────
langgraph>=0.4.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
# langchain-anthropic>=0.3.0        # uncomment if using Anthropic
langchain-mcp-adapters>=0.1.0       # for MCP tool server support

# ── Databases ────────────────────────────────────────────────────
clickhouse-connect>=0.8.0
asyncpg>=0.30.0                     # PostgreSQL async driver

# ── Redis ────────────────────────────────────────────────────────
redis[hiredis]>=5.0.0

# ── Rate Limiting ────────────────────────────────────────────────
slowapi>=0.1.9

# ── Visualization ────────────────────────────────────────────────
plotly>=6.0.0
kaleido>=0.2.1                      # Plotly static image export
matplotlib>=3.9.0
pandas>=2.2.0

# ── Document Export ──────────────────────────────────────────────
openpyxl>=3.1.0                     # Excel
python-docx>=1.1.0                  # Word
reportlab>=4.2.0                    # PDF

# ── HTTP ─────────────────────────────────────────────────────────
httpx>=0.27.0
python-multipart>=0.0.12

# ── Observability ────────────────────────────────────────────────
structlog>=24.0.0

# ── Testing ──────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
